Add BRFares plugin
It doesn't do EVERYTHING in the ticket, but it's pretty close. It's
missing the ability to give a link to a textual version of the
restriction code (which would be VERY useful but I couldn't figure out
the best way to do it), and the ability to break fares up into groups if
there are too many of them. It has been tested a bit in #tim32.